requirements.txt
apache-beam==2.19.0
avro-python3==1.9.2.1
awscli==1.18.19
boto3==1.12.19
botocore==1.15.19
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
docutils==0.15.2
fastavro==0.21.24
fasteners==0.15
future==0.18.2
google-api-core==1.16.0
google-apitools==0.5.28
google-auth==1.11.2
google-cloud-bigquery==1.17.1
google-cloud-bigtable==1.0.0
google-cloud-core==1.3.0
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
google-resumable-media==0.4.1
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.27.2
hdfs==2.5.8
httplib2==0.12.0
idna==2.9
jmespath==0.9.5
mock==2.0.0
monotonic==1.5
mysql-connector==2.2.9
numpy==1.18.1
oauth2client==3.0.0
pandas==1.0.1
pbr==5.4.4
praw==6.5.1
prawcore==1.0.1
protobuf==3.11.3
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.15
typing==3.7.4.1
typing-extensions==3.7.4.1
update-checker==0.16
urllib3==1.25.8
websocket-client==0.57.0
